[eds_balancer_config_builder] rename
