+++ xds/internal/balancer/clusterresolver/balancerconfig/type.go
@@ -16,7 +16,7 @@
  *
  */
 
-package balancerconfigbuilder
+package balancerconfig
 
 import (
 	"bytes"
